requirements.txt
fastapi
jinja2
python-multipart
sqlalchemy
starlette
uvicorn
httpx
pytest
pytest-cov
langchain
langchain-community
langchain-openai
tavily-python
python-dotenv
tree-sitter
